html/browsers/browsing-the-web/back-forward-cache/eligibility/broadcast-channel.html

<!doctype html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/PrefixedLocalStorage.js"></script>
<script>
var prefixedLocalStorage;
setup(() => prefixedLocalStorage = new PrefixedLocalStorageTest());

async_test(t => {
  t.add_cleanup(() => prefixedLocalStorage.cleanup());
  prefixedLocalStorage.onSet('result', t.step_func_done(e => {
      const result = JSON.parse(e.newValue);
      assert_false(result.isBFCached, 'Should be BFCached');
  }));
  // {BackForwardCacheMetrics::NotRestoredReason::kBlocklistedFeatures},
  // {blink::scheduler::WebSchedulerTrackedFeature::kBroadcastChannel}
  const url = new URL('../resources/simple.py', location.href);
  url.searchParams.set('script', `window.bc = new BroadcastChannel('foo');`);
  window.open(prefixedLocalStorage.url(url.href),
    '_blank',
    'noopener');
}, 'BroadcastChannel is open');
</script>
